+++ src/Collection/Collection.php
@@ -115,6 +115,7 @@ class Collection implements CollectionInterface
 
     /**
      * {@inheritdoc}
+     * @param array<mixed, mixed> $params
      */
     public function __call(string $methodName, array $params): ResultInterface
     {