+++ src/Transport/Guzzle.php
@@ -80,6 +80,7 @@ class Guzzle implements TransportInterface
 
     /**
      * {@inheritdoc}
+     * @param array<mixed, mixed> $elements
      */
     public function createStream(array &$elements): StreamInterface
     {