+++ tests/Sellsy/Client/AbstractResultTest.php
@@ -54,55 +54,54 @@ abstract class AbstractResultTest extends \PHPUnit\Framework\TestCase
     public function testIsSuccess()
     {
         $result = $this->buildResultWithSuccess();
-        self::assertInternalType('bool', $result->isSuccess());
+        self::assertIsBool($result->isSuccess());
         self::assertTrue($result->isSuccess());
         $result = $this->buildResultWithError();
-        self::assertInternalType('bool', $result->isSuccess());
+        self::assertIsBool($result->isSuccess());
         self::assertFalse($result->isSuccess());
     }
 
     public function testIsError()
     {
         $result = $this->buildResultWithSuccess();
-        self::assertInternalType('bool', $result->isError());
+        self::assertIsBool($result->isError());
         self::assertFalse($result->isError());
         $result = $this->buildResultWithError();
-        self::assertInternalType('bool', $result->isError());
+        self::assertIsBool($result->isError());
         self::assertTrue($result->isError());
     }
 
     public function testGetErrorMessage()
     {
         $result = $this->buildResultWithSuccess();
-        self::assertInternalType('string', $result->getErrorMessage());
+        self::assertIsString( $result->getErrorMessage());
         self::assertEmpty($result->getErrorMessage());
         $result = $this->buildResultWithError();
-        self::assertInternalType('string', $result->getErrorMessage());
+        self::assertIsString( $result->getErrorMessage());
         self::assertEquals('fooBar', $result->getErrorMessage());
     }
 
     public function testGetRaw()
     {
         $result = $this->buildResultWithSuccess();
-        self::assertInternalType('string', $result->getRaw());
+        self::assertIsString( $result->getRaw());
         self::assertNotEmpty($result->getRaw());
     }
 
     public function testGetResponse()
     {
         $result = $this->buildResultWithSuccess();
-        self::assertInternalType('array', $result->getResponse());
+        self::assertIsArray($result->getResponse());
         self::assertNotEmpty($result->getResponse());
         $result = $this->buildResultWithError();
-        self::assertInternalType('array', $result->getResponse());
+        self::assertIsArray($result->getResponse());
         self::assertNotEmpty($result->getResponse());
     }
 
-    /**
-     * @expectedException \RuntimeException
-     */
     public function testGetResponseException()
     {
+        $this->expectException(\RuntimeException::class);
+
         $result = $this->buildResultWithNoResponse();
         $result->getResponse();
     }